Safari-611.1.10/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/css/css-writing-modes/wm-propagation-body-scroll-offset-vertical-rl-expected.txt


PASS Check initial scroll position of viewport.
PASS Scroll to scrollLeft = -1000 should be possible.
FAIL Scroll to scrollLeft = 1000 should not be possible. assert_equals: scrollLeft should be 0. expected 0 but got 1000
